[dtsh 18/36] rich: svg: command outputs redirection to SVG files
Provides a model layer between the rich library
Console.export_svg() API and DTSh command outputs redirection
to SVG files.

Although the primary rationale for this module is to support
appending a command output to an existing SVG file (and to get rid
of the macOS-like title bar generated by the rich library),
it should also help in the long term to maintain and evolve
our SVG format.

Signed-off-by: Christophe Dufaza <chris@openmarl.org>